# imported/hack-govnoprod/infra/openapijson.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Curestry API",
        "description": "AI Prompt Analysis & Optimization Platform",
        "version": "0.1.0"
    },
    "paths": {
        "/analyze/": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Analyze Prompt",
                "description": "Analyze a prompt for quality, consistency, and potential improvements.\n\nThis endpoint performs comprehensive prompt analysis including:\n- Language detection and translation\n- Format validation and markup fixing\n- Vocabulary unification\n- Contradiction detection\n- Semantic entropy analysis\n- LLM-as-judge scoring\n- Improvement patch generation\n- Clarification questions",
                "operationId": "analyze_prompt_analyze__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AnalyzeRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalyzeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/apply": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Apply Patches",
                "description": "Apply selected improvement patches to a prompt.",
                "operationId": "apply_patches_analyze_apply_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApplyPatchesRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PromptImproved"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/clarify": {
            "post": {
                "tags": [
                    "analysis"
                ],
                "summary": "Process Clarification",
                "description": "Process clarification answers and provide updated analysis.",
                "operationId": "process_clarification_analyze_clarify_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClarifyRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnalyzeResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/export/{prompt_id}.{format_type}": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Export Prompt",
                "description": "Export analyzed prompt in specified format (md or xml).",
                "operationId": "export_prompt_analyze_export__prompt_id___format_type__get",
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    },
                    {
                        "name": "format_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Format Type"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/analyze/report/{prompt_id}.json": {
            "get": {
                "tags": [
                    "analysis"
                ],
                "summary": "Download Report",
                "description": "Download full analysis report as JSON.",
                "operationId": "download_report_analyze_report__prompt_id__json_get",
                "parameters": [
                    {
                        "name": "prompt_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Prompt Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AnalyzeRequest": {
                "properties": {
                    "prompt": {
                        "$ref": "#/components/schemas/PromptInput"
                    },
                    "include_entropy": {
                        "type": "boolean",
                        "title": "Include Entropy",
                        "description": "Include semantic entropy analysis",
                        "default": true
                    },
                    "include_clarify": {
                        "type": "boolean",
                        "title": "Include Clarify",
                        "description": "Include clarification questions",
                        "default": true
                    },
                    "include_patches": {
                        "type": "boolean",
                        "title": "Include Patches",
                        "description": "Include improvement patches",
                        "default": true
                    }
                },
                "type": "object",
                "required": [
                    "prompt"
                ],
                "title": "AnalyzeRequest",
                "description": "Request to analyze a prompt."
            },
            "AnalyzeResponse": {
                "properties": {
                    "report": {
                        "$ref": "#/components/schemas/MetricReport"
                    },
                    "patches": {
                        "items": {
                            "$ref": "#/components/schemas/Patch"
                        },
                        "type": "array",
                        "title": "Patches"
                    },
                    "questions": {
                        "items": {
                            "$ref": "#/components/schemas/ClarifyQuestion"
                        },
                        "type": "array",
                        "title": "Questions"
                    }
                },
                "type": "object",
                "required": [
                    "report",
                    "patches",
                    "questions"
                ],
                "title": "AnalyzeResponse",
                "description": "Response from prompt analysis."
            },
            "ApplyPatchesRequest": {
                "properties": {
                    "prompt_id": {
                        "type": "string",
                        "title": "Prompt Id",
                        "description": "ID of the analyzed prompt"
                    },
                    "patch_ids": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Patch Ids",
                        "description": "IDs of patches to apply"
                    },
                    "apply_safe_all": {
                        "type": "boolean",
                        "title": "Apply Safe All",
                        "description": "Apply all safe patches",
                        "default": false
                    }
                },
                "type": "object",
                "required": [
                    "prompt_id",
                    "patch_ids"
                ],
                "title": "ApplyPatchesRequest",
                "description": "Request to apply specific patches."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PromptImproved": {
                "properties": {
                    "original_prompt": {
                        "type": "string",
                        "title": "Original Prompt",
                        "description": "The original prompt"
                    },
                    "improved_prompt": {
                        "type": "string",
                        "title": "Improved Prompt",
                        "description": "The improved prompt"
                    },
                    "applied_patches": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Applied Patches",
                        "description": "IDs of patches that were applied"
                    },
                    "improvement_summary": {
                        "type": "string",
                        "title": "Improvement Summary",
                        "description": "Summary of changes made"
                    },
                    "quality_gain": {
                        "type": "number",
                        "title": "Quality Gain",
                        "description": "Estimated quality improvement"
                    }
                },
                "type": "object",
                "required": [
                    "original_prompt",
                    "improved_prompt",
                    "applied_patches",
                    "improvement_summary",
                    "quality_gain"
                ],
                "title": "PromptImproved",
                "description": "Final improved prompt after applying patches."
            },
...
        }
    }
}